feat: 2.0.1 v20

+++ ej-modules/ej-rpc/src/main/java/easy4j/infra/rpc/config/E4jClientConfig.java
@@ -24,12 +24,12 @@ public class E4jClientConfig {
      * 调用超时时间
      * 调用开始 等待超过这个时间未返回那么就报错 默认30秒
      */
-    private Integer invokeTimeOutMillis = 1000 * 30;
+    private Long invokeTimeOutMillis = 1000 * 30L;
 
     /**
      * 客户端TCP三次握手阶段连接超时，控制客户端与服务器简历连接的最大耗时，避免客户端无限期等待连接建立，默认3000ms
      */
-    private Integer connectTimeOutMillis = 3000;
+    private Long connectTimeOutMillis = 3000L;
 
 
     /**
